fix layering issues
