--- src/vs/platform/extensions/common/builtinExtensions.ts
@@ -1,40 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { isWeb } from 'vs/base/common/platform';
-import { IExtension } from 'vs/platform/extensions/common/extensions';
-import { URI } from 'vs/base/common/uri';
-
-let builtinExtensions: IExtension[] = [];
-
-// Web
-if (isWeb) {
-
-	// Running out of sources
-	if (Object.keys(builtinExtensions).length === 0) {
-		// Find builtin extensions by checking for DOM
-		const builtinExtensionsElement = document.getElementById('vscode-workbench-builtin-extensions');
-		const builtinExtensionsElementAttribute = builtinExtensionsElement ? builtinExtensionsElement.getAttribute('data-settings') : undefined;
-		if (builtinExtensionsElementAttribute) {
-			try {
-				builtinExtensions = JSON.parse(builtinExtensionsElementAttribute);
-			} catch (error) { /* ignore error*/ }
-		}
-	}
-}
-
-// Unknown
-else {
-	throw new Error('Unable to resolve builtin extensions');
-}
-
-builtinExtensions = builtinExtensions.map(extension => ({
-	...extension,
-	location: URI.revive(extension.location),
-	readmeUrl: URI.revive(extension.readmeUrl),
-	changelogUrl: URI.revive(extension.changelogUrl),
-}));
-
-export default builtinExtensions;
