+++ src/vs/workbench/services/extensionManagement/browser/extensionManagementServerService.ts
@@ -13,7 +13,7 @@ import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { ILabelService } from 'vs/platform/label/common/label';
 import { isWeb } from 'vs/base/common/platform';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { WebExtensionManagementService } from 'vs/workbench/services/extensionManagement/common/webExtensionManagementService';
+import { WebExtensionManagementService } from 'vs/workbench/services/extensionManagement/browser/webExtensionManagementService';
 import { IExtension } from 'vs/platform/extensions/common/extensions';
 
 export class ExtensionManagementServerService implements IExtensionManagementServerService {
